requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

bs4==0.0.2
contractions==0.1.73
keras==3.2.1
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
streamlit==1.33.0
streamlit-option-menu==0.3.12
tensorflow==2.16.1


# altair
# pydeck
# datetime
# time
# keras==2.15.0
# tensorflow==2.15.0
# contractions
# scikit-learn
# wordcloud
# seaborn
# matplotlib
# imbalanced-learn
# regex